+++ data/src/main/java/com/microsoft/azure/kusto/data/Client.java
@@ -6,14 +6,7 @@ package com.microsoft.azure.kusto.data;
 import com.microsoft.azure.kusto.data.exceptions.DataClientException;
 import com.microsoft.azure.kusto.data.exceptions.DataServiceException;
 
-import java.io.Closeable;
-
-public interface Client extends Closeable {
-    KustoOperationResult execute(String command) throws DataServiceException, DataClientException;
-
-    KustoOperationResult execute(String database, String command) throws DataServiceException, DataClientException;
-
-    KustoOperationResult execute(String database, String command, ClientRequestProperties properties) throws DataServiceException, DataClientException;
+public interface Client {
 
     KustoOperationResult executeQuery(String command) throws DataServiceException, DataClientException;
 
@@ -27,7 +20,7 @@ public interface Client extends Closeable {
 
     KustoOperationResult executeMgmt(String database, String command, ClientRequestProperties properties) throws DataServiceException, DataClientException;
 
-    String executeToJsonResult(String database) throws DataServiceException, DataClientException;
+    String executeToJsonResult(String command) throws DataServiceException, DataClientException;
 
     String executeToJsonResult(String database, String command) throws DataServiceException, DataClientException;