+++ data/src/main/java/com/microsoft/azure/kusto/data/http/HttpClientProperties.java
@@ -2,7 +2,6 @@ package com.microsoft.azure.kusto.data.http;
 
 import com.azure.core.http.HttpClientProvider;
 import com.azure.core.http.ProxyOptions;
-import com.azure.core.http.netty.NettyAsyncHttpClientProvider;
 
 import java.time.Duration;
 
@@ -130,7 +129,7 @@ public class HttpClientProperties {
         private Integer maxKeepAliveTime = 120;
         private Integer maxConnectionsTotal = 40;
         private Integer maxConnectionsPerRoute = 40;
-        private Duration timeout = null;
+        private Duration timeout = Duration.ofMinutes(10);
         private Class<? extends HttpClientProvider> provider = null;
         private ProxyOptions proxy = null;